Make Java Smoke Test PR workflow generic * so that it can be used for every Java samples
