+++ java/leiningen/smoke_test/leiningen_test.go
@@ -1,7 +1,9 @@
-package java_test
+package leiningen_test
 
 import (
+	"flag"
 	"fmt"
+	"github.com/paketo-buildpacks/samples/tests"
 	"os"
 	"path/filepath"
 	"testing"
@@ -15,21 +17,26 @@ import (
 	. "github.com/paketo-buildpacks/occam/matchers"
 )
 
-func TestOpentelemetry(t *testing.T) {
+var builders tests.BuilderFlags
+
+func init() {
+	flag.Var(&builders, "name", "the name a builder to test with")
+}
+func TestLeiningen(t *testing.T) {
 	Expect := NewWithT(t).Expect
 
 	Expect(len(builders)).NotTo(Equal(0))
 
 	SetDefaultEventuallyTimeout(60 * time.Second)
 
-	suite := spec.New("Java - Opentelemetry", spec.Parallel(), spec.Report(report.Terminal{}))
+	suite := spec.New("Java - Leiningen", spec.Parallel(), spec.Report(report.Terminal{}))
 	for _, builder := range builders {
-		suite(fmt.Sprintf("Opentelemetry with %s builder", builder), testOpentelemetryWithBuilder(builder), spec.Sequential())
+		suite(fmt.Sprintf("Leiningen with %s builder", builder), testLeiningenWithBuilder(builder), spec.Sequential())
 	}
 	suite.Run(t)
 }
 
-func testOpentelemetryWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
+func testLeiningenWithBuilder(builder string) func(*testing.T, spec.G, spec.S) {
 	return func(t *testing.T, context spec.G, it spec.S) {
 		var (
 			Expect     = NewWithT(t).Expect
@@ -71,7 +78,7 @@ func testOpentelemetryWithBuilder(builder string) func(*testing.T, spec.G, spec.
 
 				err = docker.Image.Remove.Execute(image.ID)
 				if err != nil {
-					Expect(err).To(MatchError("failed to remove docker image: exit status 1: Error: No such image:"))
+					Expect(err).To(MatchError(ContainSubstring("failed to remove docker image: exit status 1: Error")))
 				} else {
 					Expect(err).ToNot(HaveOccurred())
 				}
@@ -79,40 +86,38 @@ func testOpentelemetryWithBuilder(builder string) func(*testing.T, spec.G, spec.
 				Expect(os.RemoveAll(source)).To(Succeed())
 			})
 
-			context("app uses opentelemetry", func() {
+			context("app uses leiningen", func() {
 				it("builds successfully", func() {
 					var err error
-					source, err = occam.Source(filepath.Join("../java", "opentelemetry"))
+					source, err = occam.Source(filepath.Join("../"))
 					Expect(err).NotTo(HaveOccurred())
 
 					var logs fmt.Stringer
 					image, logs, err = pack.Build.
-						WithPullPolicy("if-not-present").
+						WithPullPolicy("never").
 						WithEnv(map[string]string{
-							"BP_OPENTELEMETRY_ENABLED": "true",
+							"JAVA_TOOL_OPTIONS": "-XX:MaxMetaspaceSize=100M",
 						}).
 						WithBuilder(builder).
-						WithBuildpacks(
-							"paketo-buildpacks/java",
-							"gcr.io/paketo-buildpacks/opentelemetry",
-						).
 						WithGID("123").
 						Execute(name, source)
 					Expect(err).ToNot(HaveOccurred(), logs.String)
 
 					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for CA Certificates")))
 					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for BellSoft Liberica")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Gradle")))
+					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Leiningen")))
 					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Executable JAR")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for Spring Boot")))
-					Expect(logs).To(ContainLines(ContainSubstring("Paketo Buildpack for OpenTelemetry")))
 
 					container, err = docker.Container.Run.
 						WithPublish("8080").
+						WithPublishAll().
+						WithTTY().
+						WithEnv(map[string]string{
+							"JAVA_TOOL_OPTIONS": "-XX:MaxMetaspaceSize=100M"}).
 						Execute(image.ID)
 					Expect(err).NotTo(HaveOccurred())
 
-					Eventually(container).Should(Serve(ContainSubstring("UP")).OnPort(8080).WithEndpoint("/actuator/health"))
+					Eventually(container).Should(Serve(ContainSubstring("Hello World!")).OnPort(8080))
 				})
 			})
 		})
